#316 fix crazy nwb2 modification issues
